rafa/requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
torch>=2.0.0
pyyaml>=6.0
pytest>=7.4.0

# PyTorch Geometric (GNN framework)
torch-geometric>=2.4.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0

# Training utilities
tensorboard>=2.14.0
optuna>=3.0.0
